+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
-- [#267] Fix networkpolicy configuration and rbac
+- [#267] Fix networkpolicy enabled configuration and add missing rbac.
 
 ## [v3.14.0] - 2025-10-13
 ### Changed