+++ CHANGELOG.md
@@ -5,10 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#269] Functionality to create service accounts after their dependant is already installed
+  - This is necessary as service accounts will be skipped if an optional dependency is not installed.
+    If the dependency is installed later, this will ensure that the service account will be created.
+  - This works by reconciling the dependent dogu.
+
 ### Changed
 - [#265] Use `sh` instead of `bash` to support dogus which do not use our container base image with pre-installed bash.
   - This affects the startup probe which calls `doguctl` in a shell to check if the dogu is ready and
   - the support manager wich uses the shell to execute a sleep command while the dogu is in support mode.
+
 ### Fixed
 - [#267] Fix networkpolicy enabled configuration and add missing rbac.